+++ json_serializable/test/src/generic_test_input.dart
@@ -1,0 +1,28 @@
+// Copyright (c) 2018, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+part of 'json_serializable_test_input.dart';
+
+@JsonSerializable()
+class GenericClass<T extends num, S> {
+  @JsonKey(fromJson: _dataFromJson, toJson: _dataToJson)
+  Object fieldObject;
+
+  @JsonKey(fromJson: _dataFromJson, toJson: _dataToJson)
+  dynamic fieldDynamic;
+
+  @JsonKey(fromJson: _dataFromJson, toJson: _dataToJson)
+  int fieldInt;
+
+  @JsonKey(fromJson: _dataFromJson, toJson: _dataToJson)
+  T fieldT;
+
+  @JsonKey(fromJson: _dataFromJson, toJson: _dataToJson)
+  S fieldS;
+
+  GenericClass();
+}
+
+T _dataFromJson<T extends num>(Object input) => null;
+Object _dataToJson<T extends num>(T input) => null;